Rename the mismatched spellings to the settled vocabulary
One operation, one spelling, in both directions of every pair:

  - The hook cursor type Cursor becomes DecodeCursor. Reader.Cursor
    returns a ValueCursor, so the bare name named two different types;
    the hook signature is now UnmarshalSimdJSON(*DecodeCursor) error.
  - DecodeCursor.ObjectOpen and ArrayOpen become BeginObject and
    BeginArray, matching Writer and ValueCursor, and its Number becomes
    NumberText, matching Node, Value, RawValue, and ValueCursor.
  - The free function DecodeTo becomes DecodeFrom, completing the
    EncodeTo/DecodeFrom directional pair alongside Codec.DecodeFrom.
  - Writer.Float becomes Writer.Float64, width-suffixed like the other
    float helpers.
  - Node.AppendString becomes Node.AppendText: Text is the
    decoded-contents word, and the old name collided with the deleted
    escape-encoding kernel of the same spelling.
  - Codec.Decode and Codec.DecodeArray become Unmarshal and
    UnmarshalArray, restoring Marshal/Unmarshal symmetry on Codec.

Pure renames: no signature, dispatch, or behavior changes. The hook
rename does not perturb the noescape/itab machinery; both dispatch
builds are updated in lockstep.

+++ stream_writer.go
@@ -222,8 +222,8 @@ func (w *Writer) Uint(v uint64) error {
 	return w.afterValue()
 }
 
-// Float writes a number value spelled exactly like Marshal.
-func (w *Writer) Float(v float64) error {
+// Float64 writes a number value spelled exactly like Marshal.
+func (w *Writer) Float64(v float64) error {
 	if !w.beforeValue() {
 		return w.err
 	}
